requirements.txt
asgiref==3.8.1
Django==4.2.18
django-mongoengine==0.5.6
djangorestframework==3.15.2
djongo==1.3.7
mongoengine==0.29.1
mysqlclient==2.2.7
psycopg2==2.9.10
pymongo==3.11.4
pytz==2024.2
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2024.2
